Add files via upload
Now MainCore is able to map the previous channels into the new ones, so that CTF output can be adequately read. In order to store that map for its later use, Channelsmap.txt is created. 

As of this moment, the script cannot erase "Channelsmap.txt" -I am still figuring why-  if it is already present in the working directory, so it has to be manually deleted (or be moved to a safe directory if one wants to use it later should he/she re-run MainCore for another case